moved add location terms to seperate util file for now
